+++ web-developer-challenge/src/components/CreatePost/CreateStyles.ts
